% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/third-eye.r
\name{cyclops}
\alias{cyclops}
\title{Retrieve TV News Archive chyrons from the Internet Archive's Third Eye project}
\usage{
cyclops(chyron_day = Sys.Date() - 1, cleaned = TRUE)
}
\arguments{
\item{chyron_day}{archive day (\code{Date} or \code{character}; if \code{character} should be
in \code{YYYY-mm-dd} format)}

\item{cleaned}{logical, default \code{TRUE}. The "raw feed" option provides all of the
OCR'ed text from chyrons at the rate of approximately one entry per second.
The "clean feed" download provides the data feed that fuels the Third Eye
Twitter bots; this has been filtered to find the most representative,
clearest chyrons from a 60-second period, with no more than one entry/tweet per
minute (though the duration may be shorter than 60 seconds.) The clean feed
relies on algorithms that are a work in progress.}
}
\value{
\code{NULL} on irrecoverable errors, otherwise a data frame with five columns:
\itemize{
\item \code{ts} (\code{POSIXct}) chyron timestamp
\item \code{channel} (\code{character}) news channel the chyron appeared on
\item \code{duration} (\code{integer}) see Description
\item \code{details} (\code{character}) Internet Archive details path
\item \code{text} (\code{character}) the cyhron text
}
}
\description{
The TV News Archive's Third Eye project captures the chyrons–or narrative text–that appear on the lower third of TV news screens and turns them into downloadable data and a Twitter feed for research, journalism, online tools, and other projects. At project launch (September 2017) we are collecting chyrons from BBC News, CNN, Fox News, and MSNBC–more than four million collected over just two weeks. Chyrons have public value because:
\itemize{
\item Breaking news often appears on chyrons before TV newscasters begin reporting or video is available, whether it's a hurricane or a breaking political story.
\item Which chyrons a TV news network chooses to display can reveal editorial decisions that can inform public understanding of how news is filtered for different audiences.
\item Providing chyrons as data–and also on Twitter–in near real-time can serve as a alert system, showing how TV news stations are reporting the news. Often the chyrons are ahead of the general conversation on Twitter.
}
}
\details{
Some notes on the data
\itemize{
\item chyrons are derived in near real-time from the TV News Archive's collection of TV news. The constantly updating public collection contains 1.4 million TV news shows, some dating back to 2009.
\item At launch, Third Eye captures four TV cable news channels: BBC News, CNN, Fox News, and MSNBC.
\item Data can be affected by temporary collection outages, which typically can last minutes or hours, but rarely more.
\item Dates/times are in UTC (Coordinated Universal Time).
\item Because the size of the raw data is so large (about 20 megabytes per day), results are limited to seven days per request.
\item Raw data collection began on August 25, 2017; the clean feed begins on September 7, 2017.
\item "\code{duration}" column is in seconds–the amount of time that particular chyron appeared on the screen.
}
}
\note{
It is \emph{highly} recommended that you use the "clean" feed unless you're researching
how to correct text. This package does it's best to read in the raw feed but
it often contains embedded nulls and non-standard text encodings which
make it difficult to process.
}
