% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-tending-range.r
\name{iatv_top_trending}
\alias{iatv_top_trending}
\title{Top Trending Topics (Internet Archive TV Archive)}
\usage{
iatv_top_trending(from, to, .progress = interactive())
}
\arguments{
\item{from, to}{start and end date/time ranges (will auto-convert if properly formatted strings)}

\item{.progress}{show a progress bar? Defaukts to \code{TRUE} if in an interactive session.}
}
\description{
Provide start & end times in current time zone and this function will generate
the proper "every 15-second" values, convert them to GMT values and issue the queries,
returning a nested data frame of results. If you want more control, use \code{\link[=top_trending]{top_trending()}}.
}
\details{
GDELT now generates a snapshot every 15 minutes that records all of the "top trending"
tables into a single archive enabling users to ook back over time at what was trending
in 15 minute increments historically back to midnight on 2017-09-07.

Note that the archives are generated every 15 minutes based on the television shows that
have completed processing at that time. It can take several hours for a show to be fully
processed by the Internet Archive and available for processing, thus the presence/absence
of a topic in these files should not be used to date it precisely to that 15 minute mark,
but rather as a rough temporal indicator of what topics were trending up/down in that
general time frame. For precise timelines, you should take a topic from this archive and
run a search on it using the main Television Explorer interface, select a timeframe of
72 hours and use the resulting timeline to precisely date the topic's coverage (since
the Explorer timeline is based on the broadcast timestamp of the show, even if it is
processed hours later).
}
\note{
The times are auto-converted to GMT
}
\examples{
top_trending("2017-09-08 18:00", "2017-09-09 06:00")
}
