% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-chyrons.r
\name{list_chyrons}
\alias{list_chyrons}
\title{Retrieve Third Eye chyron index}
\usage{
list_chyrons()
}
\value{
data frame with three columns:
\itemize{
\item \code{ts} (\code{POSIXct}) chyron timestamp
\item \code{type} (\code{character}) \code{raw} or \code{cleaned}
\item \code{size} (\code{numeric}) size of the feed file in bytes
}
}
\description{
Returns a data frame with available chyron dates & selected metadata.
}
