% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.r
\name{list_networks}
\alias{list_networks}
\title{Helper function to identify station/network keyword and corpus date range for said market}
\usage{
list_networks(widget = interactive())
}
\arguments{
\item{widget}{if `TRUE` then an HTML widget will be displayed to make it easier to
sift through stations/networks}
}
\value{
data frame
}
\description{
The \code{filter_network} of \code{query_tv()} is picky so this helps you idenitify the
keyword to use for the particular network/station.
}
\details{
The list also shows the date ranges available for the captions, so you can use that as
a guide when picking dates.

In interactive mode it uses \code{DT::datatable()}. You can force it to just display to
the console by passing in \code{widget=FALSE}
}
\examples{
list_networks() # widget
print(list_networks(FALSE)) # no widget
}
