% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsflash-package.R
\docType{package}
\name{newsflash}
\alias{newsflash}
\alias{newsflash-package}
\title{Tools to Work with the Internet Archive and GDELT Television Explorer}
\description{
Tools to Work with the Internet Archive and GDELT Television Explorer
}
\author{
Bob Rudis (bob@rud.is)
}

