% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ndf.r
\name{print.newsflash}
\alias{print.newsflash}
\title{Displays a summary of the query results}
\usage{
\method{print}{newsflash}(x)
}
\arguments{
\item{x}{A \code{newsflash} object}
}
\description{
Use \code{str()} for an object-level overview.
}
