% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-trending.r
\name{top_trending}
\alias{top_trending}
\title{Top Trending Tables}
\usage{
top_trending(ymd = Sys.Date(), hour = 0:24, minute = c(0, 15, 30, 45))
}
\arguments{
\item{ymd}{a \code{Date} object or a character string in \code{YYYYMMDD} or \code{YYYY-MM-DD} format.
(Defaults to today).}

\item{hour}{\code{0} to \code{23} (Defaults to \code{0})}

\item{minute}{one of \code{0}, \code{15}, \code{30}, or \code{45}. (Defaults to \code{0})}
}
\value{
\code{list} with
\itemize{
\item \code{DateGenerated} (character vector)
\item \code{OverallTrendigTopics} (character vector)
\item \code{StationTrendingTopics} (data frame with a \code{Station} column and a \code{list} column of \code{Topics})
\item \code{StationTopTopics} (same structure as \code{StationTrendingTopics})
\item \code{OverallTrendingPhrases}
}
}
\description{
GDELT now generates a snapshot every 15 minutes that records all of the "top trending"
tables into a single archive enabling users to ook back over time at what was trending
in 15 minute increments historically back to midnight on 2017-09-07.
}
\details{
Note that the archives are generated every 15 minutes based on the television shows that
have completed processing at that time. It can take several hours for a show to be fully
processed by the Internet Archive and available for processing, thus the presence/absence
of a topic in these files should not be used to date it precisely to that 15 minute mark,
but rather as a rough temporal indicator of what topics were trending up/down in that
general time frame. For precise timelines, you should take a topic from this archive and
run a search on it using the main Television Explorer interface, select a timeframe of
72 hours and use the resulting timeline to precisely date the topic's coverage (since
the Explorer timeline is based on the broadcast timestamp of the show, even if it is
processed hours later).

\strong{The time is expected to be GMT for the API and this function will eventually be
modified to auto-convert local timezone to GMT with a parameter.}
}
\note{
If an error occurred with the API or transmission, \code{NULL} is returned.
}
\examples{
top_trending(hour=14, minute=30)
}
