% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-clip.r
\name{view_clip}
\alias{view_clip}
\title{View news segment clips from chyron details frament}
\usage{
view_clip(details_fragment)
}
\arguments{
\item{details_fragment}{a single entry from the `details` column of a data
frame created by [cyclops()]. It will look something like:
`FOXNEWSW_20170919_100000_FOX__Friends/start/792`}
}
\description{
Opens a browser to show the a clip of the scraped segment
}
