% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-cloud.R
\name{word_cloud}
\alias{word_cloud}
\title{Retrieve top words that appear most frequently in clips matching your search}
\usage{
word_cloud(query, start_date = NULL, end_date = NULL)
}
\arguments{
\item{query}{query string in GDELT format. See \code{QUERY} in https://blog.gdeltproject.org/gdelt-2-0-television-api-debuts/
for details; use \code{\link[=list_networks]{list_networks()}} to obtain valid station/network identifiers}

\item{start_date, end_date}{start/end dates. Leaving both \code{NULL} searches all archive history.
Leaving just \code{start_date} \code{NULL} sets the start date to July 2009. Leaving just \code{end_date}
\code{NULL} sets the end date to today.}
}
\description{
The API takes the 200 most relevant clips that match your search and returns the
terms for a "word cloud" of up to the top 200 most frequent words that appeared in
those clips (common stop words are automatically removed). This is a powerful way
of understanding the topics and words dominating the relevant coverage and
suggesting additional contextual search terms to narrow or evolve your search.
Note that if there are too few matching clips for your query, the word cloud may
be blank.
}
