% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsflash.r
\name{query_tv}
\alias{query_tv}
\title{Issue a query to the TV Explorer}
\usage{
query_tv(primary_keyword, context_keywords = NULL,
  filter_network = "NATIONAL", start_date = as.Date("2016-10-01"),
  end_date = as.Date("2016-12-31"))
}
\arguments{
\item{primary_keyword}{primary keyword}

\item{context_keywords}{context keywords}

\item{filter_network}{filter by network. One of "NATIONAL", "ALJAZAM", "BLOOMBERG", "CNBC", "CNN", "FBC", "FOXNEWSW",
"MSNBC", "INTERNATIONAL", "BBCNEWSSEG", "AFFNETALL", "AFFNET_ABC",
"AFFNET_CBS", "AFFNET_FOX", "AFFNET_MYTV", "AFFNET_NBC", "AFFNET_PBS",
"AFFMARKALL", "AFFMARKET_Boston", "AFFMARKET_Cedar Rapids", "AFFMARKET_Charlotte",
"AFFMARKET_Cincinnati", "AFFMARKET_Cleveland", "AFFMARKET_Colorado Springs",
"AFFMARKET_Columbia", "AFFMARKET_Dakota Dunes SD", "AFFMARKET_Daytona Beach",
"AFFMARKET_Denver", "AFFMARKET_Des Moines", "AFFMARKET_Durham",
"AFFMARKET_Goldsboro", "AFFMARKET_Greenville", "AFFMARKET_Hampton",
"AFFMARKET_Las Vegas", "AFFMARKET_Lynchburg", "AFFMARKET_Miami",
"AFFMARKET_Newport KY", "AFFMARKET_Norfolk", "AFFMARKET_Orlando",
"AFFMARKET_Philadelphia", "AFFMARKET_Portsmouth", "AFFMARKET_Pueblo",
"AFFMARKET_Raleigh", "AFFMARKET_Reno", "AFFMARKET_Roanoke", "AFFMARKET_San Francisco",
"AFFMARKET_Shaker Heights", "AFFMARKET_Sioux City", "AFFMARKET_St. Petersburg",
"AFFMARKET_Tampa", "AFFMARKET_Virginia Beach", "AFFMARKET_Washington DC",
"AFFMARKET_Waterloo"}

\item{start_date, end_date}{\code{Date} object for start/end of search. Defaults Oct-2016 to Dec-2016}
}
\description{
Successful queries (API errors tend to be date-range related but there
isn't enough to go on in the API response to provide better diagnostics)
result in a list of 4 tibbles.
}
\details{
The first is \code{query_details} which spits back what you put in.

The second is \code{timeline} which is a long tibble of the number of times
the keywords combo was on a particulare station.

The third is \code{station_histogram} which is what it says it is.

The fourth and last is \code{top_matches} which is also what it says it is except
that it includes juicy metadata with the Internet Archive preview URL,
thumbnail URL, caption snippet and more.

See the Reference URL for parameter info
}
\references{
\url{http://television.gdeltproject.org/cgi-bin/iatv_ftxtsearch/iatv_ftxtsearch}
}

