% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsflash.r
\name{query_tv}
\alias{query_tv}
\title{Issue a query to the TV Explorer}
\usage{
query_tv(primary_keyword, context_keywords = NULL,
  filter_network = "NATIONAL", start_date = as.Date("2016-10-01"),
  end_date = as.Date("2016-12-31"))
}
\arguments{
\item{primary_keyword}{primary keyword}

\item{context_keywords}{context keywords (optional, but see web site for details on
why this is a useful parameter)}

\item{filter_network}{filter by network. Use \code{list_networks()} to see valid values
and to get a hint about valid date ranges for \code{start_date} and \code{end_date}.
Defaults to \code{NATIONAL}.}

\item{start_date, end_date}{\code{Date} object for start/end of search. Defaults to the
last quarter of 2016: Oct-2016 to Dec-2016}
}
\description{
Successful queries (API errors tend to be date-range related but there
isn't enough to go on in the API response to provide better diagnostics)
result in a list of 4 tidy tibbles:
}
\details{
\describe{
\item{\code{query_details}}{spits back what you put in.}
\item{\code{timeline}}{a long tibble of the number of times
the keywords combo was on a particulare station.}
\item{\code{station_histogram}}{what it says it is.}
\item{\code{top_matches}}{also what it says it is except
that it includes juicy metadata with the Internet Archive preview URL,
thumbnail URL, \emph{caption text snippet} and more.}
}

Queries resulting in no matches will display a message and return \code{NULL} invisibly.

See the Reference URL for parameter info
}
\examples{
query_tv("terror", "isis")
query_tv("trump")
}
\references{
\url{http://television.gdeltproject.org/cgi-bin/iatv_ftxtsearch/iatv_ftxtsearch}
}

