% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_text.r
\name{top_text}
\alias{top_text}
\title{Helper function to extract the text snippets from the top matches}
\usage{
top_text(x)
}
\arguments{
\item{x}{query result object from calling \code{query_tv}}
}
\description{
Helper function to extract the text snippets from the top matches
}

